IMP: Changed some small wordy/punctuation bits
I like this a lot, not much from a substantive perspective I have to
offer. Made me stop and think about our current process, where our SC
is, how easy (or not!) it is for other teams to contribute, what our
process is and how it's usually not documented but tribal knowledge, and
how those components can impact one another.

What might be nice as a follow up would be to make some suggestions on
how you could make it easier for onboarding, cross-team collaboration,
sharing best practices, etc. Or, how could a team go from maybe your
example 1 to example 2. Not sure if you wanted to get into that or not.